fix resize handle and improve editor workspace defaults
Because of changes to how the editor is layed out it was possible
to drag the handle further than the width of the screen. This
prevents you from doing that unless you have significantly decreased
the parent width of the webpage. This also fixes the height of the
editor and the handle when the parent width of the webpage is low.