Fixed broken links and /newline issues
I opted to fix the links by adding `../` instead of removing `md/`. This is because I figured for maintainability, it's much easier to do a find/replace on `"../md` than it is on `"`.

I also got rid of a few extraneous bullets and fixed some newline issues which arose because the trailing `>` on the Markdown HTML needs a blank newline or else it causes the subsequent character to get garbled.